tests/submissions/anssik/battery-discharging.html

<!DOCTYPE html>
<html>
  <head>
  <title>Battery Status API Test Suite</title>
  <script src="http://www.w3c-test.org/resources/testharness.js"></script>
  <link rel='stylesheet' href='http://w3c-test.org/resources/testharness.css' media='all'/>
  </head>
  <body>
    <h1>Description</h1>
    <p>
      The battery is discharging.
    </p>
    <h2>Preconditions</h2>
    <ol>
      <li>
        The device is unplugged from the charger before this test case is run.
      </li>
      <li>
        The battery must neither be empty or full, nor reach empty or full capacity during the test.
      </li>
    </ol>
    <div id="log"></div>
    <script>
    (function() {

      test(function() {
        assert_false(navigator.battery.charging);
      }, 'The charging attribute must be set to false if the battery is discharging.');
      
      test(function() {
        assert_true(navigator.battery.chargingTime === Infinity);
      }, 'The chargingTime attribute must be set to the value positive Infinity if the battery is discharging');
      
      test(function() {
        assert_true(navigator.battery.dischargingTime < Infinity);
      }, 'The dischargingTime attribute must be set to the time remaining in ' +
      'seconds until the system\'s battery is completely discharged.');
      
      test(function() {
        assert_true(0 <= navigator.battery.level && navigator.battery.level <= 1.0);
      }, 'The level attribute must be set to the current battery level scaled from 0 to 1.0');
      
    })();
    </script>
  </body>
</html>
